pyproject.toml
[project]
name = "due"
version = "0.1.0"
description = "due.moe"
authors = [
    { name = "Fuwn", email = "contact@fuwn.me" }
]
dependencies = [
    "flask>=2.3.2",
    "requests>=2.31.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "joblib>=1.3.1",
    "Flask-Caching>=2.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
